chore: migrate PreCheckExistence hook to validator
