+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/config/TrackerValidationConfig.java
@@ -34,7 +34,6 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.hisp.dhis.tracker.validation.TrackerValidationHook;
-import org.hisp.dhis.tracker.validation.hooks.PreCheckExistenceValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.PreCheckUidValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.RepeatedEventsValidationHook;
 import org.springframework.context.annotation.Bean;
@@ -72,7 +71,6 @@ public class TrackerValidationConfig
     public List<TrackerValidationHook> validationHooks()
     {
         return getHookByClass( ImmutableList.of( PreCheckUidValidationHook.class,
-            PreCheckExistenceValidationHook.class,
 
             /*
              * NB! This hook must be run after all the Event validations,
