+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/hooks/RelationshipPreCheckExistenceValidator.java
@@ -27,114 +27,26 @@
  */
 package org.hisp.dhis.tracker.validation.hooks;
 
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1002;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1030;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1032;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1063;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1080;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1081;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1082;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1113;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1114;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E4015;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E4016;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E4017;
 
-import org.hisp.dhis.program.ProgramInstance;
-import org.hisp.dhis.program.ProgramStageInstance;
-import org.hisp.dhis.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.tracker.TrackerImportStrategy;
 import org.hisp.dhis.tracker.bundle.TrackerBundle;
-import org.hisp.dhis.tracker.domain.Enrollment;
-import org.hisp.dhis.tracker.domain.Event;
 import org.hisp.dhis.tracker.domain.Relationship;
-import org.hisp.dhis.tracker.domain.TrackedEntity;
-import org.hisp.dhis.tracker.validation.TrackerValidationHook;
 import org.hisp.dhis.tracker.validation.ValidationErrorReporter;
+import org.hisp.dhis.tracker.validation.Validator;
 import org.springframework.stereotype.Component;
 
 /**
  * @author Morten Svanæs <msvanaes@dhis2.org>
  */
 @Component
-public class PreCheckExistenceValidationHook
-    implements TrackerValidationHook
+public class RelationshipPreCheckExistenceValidator
+    implements Validator<Relationship>
 {
     @Override
-    public void validateTrackedEntity( ValidationErrorReporter reporter, TrackerBundle bundle,
-        TrackedEntity trackedEntity )
-    {
-        TrackerImportStrategy importStrategy = bundle.getStrategy( trackedEntity );
-
-        TrackedEntityInstance existingTe = bundle.getPreheat().getTrackedEntity( trackedEntity.getTrackedEntity() );
-
-        // If the tracked entity is soft-deleted no operation is allowed
-        if ( existingTe != null && existingTe.isDeleted() )
-        {
-            reporter.addError( trackedEntity, E1114, trackedEntity.getTrackedEntity() );
-            return;
-        }
-
-        if ( existingTe != null && importStrategy.isCreate() )
-        {
-            reporter.addError( trackedEntity, E1002, trackedEntity.getTrackedEntity() );
-        }
-        else if ( existingTe == null && importStrategy.isUpdateOrDelete() )
-        {
-            reporter.addError( trackedEntity, E1063, trackedEntity.getTrackedEntity() );
-        }
-    }
-
-    @Override
-    public void validateEnrollment( ValidationErrorReporter reporter, TrackerBundle bundle, Enrollment enrollment )
-    {
-        TrackerImportStrategy importStrategy = bundle.getStrategy( enrollment );
-
-        ProgramInstance existingPi = bundle.getPreheat().getEnrollment( enrollment.getEnrollment() );
-
-        // If the tracked entity is soft-deleted no operation is allowed
-        if ( existingPi != null && existingPi.isDeleted() )
-        {
-            reporter.addError( enrollment, E1113, enrollment.getEnrollment() );
-            return;
-        }
-
-        if ( existingPi != null && importStrategy.isCreate() )
-        {
-            reporter.addError( enrollment, E1080, enrollment.getEnrollment() );
-        }
-        else if ( existingPi == null && importStrategy.isUpdateOrDelete() )
-        {
-            reporter.addError( enrollment, E1081, enrollment.getEnrollment() );
-        }
-    }
-
-    @Override
-    public void validateEvent( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
-    {
-        TrackerImportStrategy importStrategy = bundle.getStrategy( event );
-
-        ProgramStageInstance existingPsi = bundle.getPreheat().getEvent( event.getEvent() );
-
-        // If the event is soft-deleted no operation is allowed
-        if ( existingPsi != null && existingPsi.isDeleted() )
-        {
-            reporter.addError( event, E1082, event.getEvent() );
-            return;
-        }
-
-        if ( existingPsi != null && importStrategy.isCreate() )
-        {
-            reporter.addError( event, E1030, event.getEvent() );
-        }
-        else if ( existingPsi == null && importStrategy.isUpdateOrDelete() )
-        {
-            reporter.addError( event, E1032, event.getEvent() );
-        }
-    }
-
-    @Override
-    public void validateRelationship( ValidationErrorReporter reporter, TrackerBundle bundle,
+    public void validate( ValidationErrorReporter reporter, TrackerBundle bundle,
         Relationship relationship )
     {
